- do not restrict sample rate to be integer!! (important only for isruc SaO2 channel. - sort files with natural order to get correct psg and annot together in ISRUC - some code cleaning/transformations - add preprocessing for ISRUC for resorting

+++ psg_processing/file_handlers/edf_handler.py
@@ -55,8 +55,8 @@ class EDFHandler(FileHandler):
             self.logger.info(f"Select channel samples: {ch_samples[select_ch_idx]}")
             ch_freq = psg_f.getSampleFrequencies()
 
-            sampling_rate = int(ch_freq[select_ch_idx])
-            n_epoch_samples = int(epoch_duration * sampling_rate)
+            sampling_rate = ch_freq[select_ch_idx]
+            n_epoch_samples = epoch_duration * sampling_rate
             signal = psg_f.readSignal(select_ch_idx)
             
             unit = psg_f.getPhysicalDimension(select_ch_idx)
